refactor: instantiate renderer-compiler
